Move getECCTemplate
